Java: OpenTelemetry update tests
