+++ java/client/src/main/java/glide/managers/CommandManager.java
@@ -584,9 +584,18 @@ public class CommandManager {
         populateCommandWithArgs(arguments, commandBuilder);
 
         long spanPtr = 0;
-        if (OpenTelemetry.isInitialized() && OpenTelemetry.shouldSample()) {
-            // Create OpenTelemetry span
-            spanPtr = OpenTelemetryResolver.createLeakedOtelSpan(requestType.name());
+        try {
+            System.out.println("IsInitialized=" + OpenTelemetry.isInitialized());
+            System.out.println("should sample=" + OpenTelemetry.shouldSample());
+            if (OpenTelemetry.isInitialized() && OpenTelemetry.shouldSample()) {
+                // Create OpenTelemetry span
+                System.out.println("Open telemetry already initialized");
+                spanPtr = OpenTelemetryResolver.createLeakedOtelSpan(requestType.name());
+            } else {
+                System.out.println("Open telemetry not initialized");
+            }
+        } catch (Exception e) {
+            System.out.println("exception in prepare command="+e.getMessage());
         }
 
         CommandRequest.Builder builder =
